Added the player names to the event reporting popup player picker.
Updated the event reporting popup to take in information that will be relevant to update the brackets.

Made players in the set class nullable because startgg has null in places where players have not yet been determined.

Added code to swap event groups; untested.

Added a set id to the Match class, so the scheduler can update the brackets and generate new matches.

Debugged the event bracket group to find out that the whole group needs to be treated as a single bracket.